requirements.txt
Django~=5.1.4
pillow
django-bootstrap5
mysqlclient
django-filter